+++ YapDatabase/Internal/YapMemoryTable.m
@@ -148,7 +148,7 @@
 				__unsafe_unretained YapMemoryTableValue *prvValue = nil;
 				__unsafe_unretained YapMemoryTableValue *value = [dict objectForKey:key];
 				
-				while (value && value->snapshot >= minSnapshot)
+				while (value && value->snapshot >= (uint64_t)minSnapshot)
 				{
 					if (hasObject == NO)
 						hasObject = (value->object != nil);
@@ -217,7 +217,7 @@
 		{
 			__unsafe_unretained YapMemoryTableValue *value = [dict objectForKey:key];
 			
-			if (value && value->snapshot == snapshot)
+			if (value && value->snapshot == (uint64_t)snapshot)
 			{
 				if (value->olderValue == nil)
 				{
